Move gamestate struct into seprate file
Next step is move get/set function in checkers game
