+++ checkers-ai/makefile
@@ -13,6 +13,7 @@ NETWORK=network
 NETWORK_READER=network-file-reader
 NETWORK_WRITER=network-file-writer
 SEARCH=search
+GAME_STATE=game-state
 MAIN=main
 
 GENERATOR_OBJ=$(BUILD_DIR)/$(GENERATOR).o
@@ -27,16 +28,17 @@ NETWORK_OBJ=$(BUILD_DIR)/$(NETWORK).o
 NETWORK_READER_OBJ=$(BUILD_DIR)/$(NETWORK_READER).o
 NETWORK_WRITER_OBJ=$(BUILD_DIR)/$(NETWORK_WRITER).o
 SEARCH_OBJ=$(BUILD_DIR)/$(SEARCH).o
+GAME_STATE_OBJ=$(BUILD_DIR)/$(GAME_STATE).o
 
 OBJ_FILES=$(GENERATOR_OBJ) $(CONVERTER_OBJ) $(GAME_OBJ) $(PLAYER_OBJ) $(UTILS_OBJ) \
 		  $(BOARD_OBJ) $(SEEDER_OBJ) \
 		  $(NETWORK_OBJ) $(NETWORK_READER_OBJ) $(NETWORK_WRITER_OBJ) \
-		  $(SEARCH_OBJ)
+		  $(SEARCH_OBJ) $(GAME_STATE_OBJ)
 
 CPP_FILES=$(GENERATOR).cpp $(CONVERTER).cpp $(GAME).cpp $(PLAYER).cpp \
 		  $(UTILS).cpp $(BOARD).cpp $(SEEDER).cpp                     \
 		  $(NETWORK).cpp $(NETWORK_WRITER).cpp $(NETWORK_READER).cpp \
-		  $(SEARCH).cpp
+		  $(SEARCH).cpp $(GAME_STATE).cpp
 
 OBJ_FILES_WITH_MAIN=$(OBJ_FILES) $(MAIN).o
 
@@ -82,6 +84,9 @@ $(NETWORK_WRITER_OBJ): $(NETWORK_WRITER).cpp
 
 $(SEARCH_OBJ): $(SEARCH).cpp
 	$(CMP) -o $(SEARCH_OBJ) -c $(SEARCH).cpp
+
+$(GAME_STATE_OBJ): $(GAME_STATE).cpp
+	$(CMP) -o $(GAME_STATE_OBJ) -c $(GAME_STATE).cpp
 
 
 PYVERSION=python3.5m
